+++ _src/tests/services/test/buffer.executor.ts
@@ -6,8 +6,8 @@ export class TestBufferExecutor implements ServiceExecutor<TestBufferRequest, Te
 	static serviceName: string = 'NaniumTest:test/buffer';
 
 	async execute(request: TestBufferRequest, executionContext: ServiceRequestContext): Promise<TestBufferResponse> {
-		request.body.buffer1?.write('*');
-		request.body.buffer2?.write('*');
+		request.body.buffer1?.write(new TextEncoder().encode('*'));
+		request.body.buffer2?.write(new TextEncoder().encode('*'));
 		return new TestBufferResponse({
 			id: request.body.id,
 			text1: await request.body.buffer1?.asString(),